% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq-map.R
\name{eq_map}
\alias{eq_map}
\title{Leaflet map of earthquakes
Creates a map of selected earthquakes by country and time.}
\usage{
eq_map(df, annotation_col)
}
\arguments{
\item{df}{A data frame containing cleaned NOAA earthquake data}

\item{annotation_col}{A character. Name of a descrptive vector in the data.}
}
\value{
A leaflet map with earthquakes and annotations.
}
\description{
Leaflet map of earthquakes
Creates a map of selected earthquakes by country and time.
}
\examples{
\dontrun{
eq_map(df, annotation_col = "location_name")
}
}
